Some refactoring of cockpit classes
