+++ engine/CMakeLists.txt
@@ -139,6 +139,7 @@ INCLUDE_DIRECTORIES(
     ${Vega_Strike_SOURCE_DIR}/src/cmd
     ${Vega_Strike_SOURCE_DIR}/src/damage
     ${Vega_Strike_SOURCE_DIR}/src/resource
+    ${Vega_Strike_SOURCE_DIR}/src/cockpit
     ${Vega_Strike_BINARY_DIR}
 )
 ELSE (MSVC)
@@ -147,6 +148,7 @@ INCLUDE_DIRECTORIES(
     ${Vega_Strike_SOURCE_DIR}/src/cmd
     ${Vega_Strike_SOURCE_DIR}/src/damage
     ${Vega_Strike_SOURCE_DIR}/src/resource
+    ${Vega_Strike_SOURCE_DIR}/src/cockpit
     ${Vega_Strike_BINARY_DIR}
     /usr/include/harfbuzz/
 )
@@ -778,6 +780,11 @@ SET(LIBRESOURCE
     src/resource/resource.cpp
     )
 
+SET(LIBCOCKPIT
+    src/cockpit/cockpit_audio.cpp
+    src/cockpit/cockpit_gfx.cpp
+    )
+
 SET(LIBGUI_SOURCES
     src/gui/button.cpp
     src/gui/control.cpp
@@ -1172,6 +1179,7 @@ ADD_LIBRARY(vegastrike-engine_com
     ${LIBCONFIG}
     ${LIBDAMAGE}
     ${LIBRESOURCE}
+    ${LIBCOCKPIT}
     ${LIBAI_SOURCES}
     ${LIBCMD_SOURCES}
     ${LIBNET_SOURCES}
