+++ engine/src/cmd/ai/firekeyboard.cpp
@@ -57,6 +57,7 @@
 #include "mount_size.h"
 #include "weapon_info.h"
 #include "vs_logging.h"
+#include "cockpit/cockpit_audio.h"
 
 extern bool toggle_pause();
