+++ engine/src/gfx/cockpit_generic.h
@@ -31,6 +31,7 @@
 #include <vector>
 #include "gfx/vec.h"
 #include "vsfilesystem.h"
+#include "cockpit/cockpit_events.h"
 //using namespace XMLSupport;  DONT DO THIS IN HEADERS
 enum VIEWSTYLE
 {
@@ -62,84 +63,7 @@ class Cockpit
 {
 public:
     ///cockpit events.
-    enum EVENTID {
-        EVENTID_FIRST,
 
-        /// Warp is possible at this time
-        WARP_READY = EVENTID_FIRST,
-
-        /// Warp ceased to be possible at this time
-        WARP_UNREADY,
-
-        /// Warp started
-        WARP_ENGAGED,
-
-        /// Warp stopped
-        WARP_DISENGAGED,
-
-        /// Asap autopilot engaged
-        ASAP_ENGAGED,
-
-        /// Asap autopilot disengaged
-        ASAP_DISENGAGED,
-
-        /// Asap docking available
-        ASAP_DOCKING_AVAILABLE,
-
-        /// Asap docking engaged
-        ASAP_DOCKING_ENGAGED,
-
-        /// Asap docking disengaged
-        ASAP_DOCKING_DISENGAGED,
-
-        /// Docking ready
-        DOCK_AVAILABLE,
-
-        /// Docking no longer ready
-        DOCK_UNAVAILABLE,
-
-        /// Docking triggered yet not ready
-        DOCK_FAILED,
-
-        /// Jump ready
-        JUMP_AVAILABLE,
-
-        /// Jump no longer ready
-        JUMP_UNAVAILABLE,
-
-        /// Jump triggered yet not ready
-        JUMP_FAILED,
-
-        /// Weapons lock warning active
-        LOCK_WARNING,
-
-        /// Missile lock warning acgive
-        MISSILELOCK_WARNING,
-
-        /// Eject light
-        EJECT_WARNING,
-
-        /// Governor enabled
-        FLIGHT_COMPUTER_ENABLED,
-
-        /// Governor disabled
-        FLIGHT_COMPUTER_DISABLED,
-
-        /// Warp loop, warp speed 0 (+1 = warp 1, +2 = warp 2, etc...)
-        WARP_LOOP0,
-        WARP_LOOP9 = WARP_LOOP0 + 9,
-        /// Last warp level
-        WRAP_LOOPLAST = WARP_LOOP9,
-
-        /// Warp threshold, warp speed 0 (+1 = warp 1, +2 = warp 2, etc...)
-        WARP_SKIP0,
-        WARP_SKIP9 = WARP_SKIP0 + 9,
-        /// Last warp level
-        WRAP_SKIPLAST = WARP_SKIP9,
-
-        /// Just after all valid values
-        NUM_EVENTS
-    };
 
 
 protected:
@@ -190,13 +114,6 @@ protected:
 
     ///draws the navigation symbol around targetted location
     virtual void DrawNavigationSymbol( const Vector &loc, const Vector &p, const Vector &q, float size ) {}
-
-    ///draws the target box around targetted unit
-    virtual float computeLockingSymbol( Unit *par )
-    {
-        return 1;
-    }
-    virtual void DrawTargetBox() {}
 
     ///draws the target box around all units
     virtual void DrawTargetBoxes(const Radar::Sensor&) {}
@@ -368,13 +285,13 @@ public:
      * Retrieves the sound associated to the given event.
      * Returns NULL if no sound has been associated
      */
-    SoundContainer* GetSoundForEvent(EVENTID eventId) const;
+    SoundContainer* GetSoundForEvent(CockpitEvent eventId) const;
 
     /**
      * Sets the sound associated to the given event to match the given specs
      * (the actual container will be a different, implementation-specific one)
      */
-    void SetSoundForEvent(EVENTID eventId, const SoundContainer &soundSpecs);
+    void SetSoundForEvent(CockpitEvent eventId, const SoundContainer &soundSpecs);
 
 
     class SaveGame*savegame;
